test: extract Groonga helper
